Se separo el componente levelBar y se parametrizo para ser usado individualmente
